# docs/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: WeddingOS Unified API
  version: 1.0.0
  description: |
    Unified OpenAPI 3.0 specification for the WeddingOS microservices platform.

    This document was assembled from the actual service route files and related
    Zod / Prisma / Pydantic definitions in the monorepo. Paths are documented as
    gateway-facing routes using the service prefixes requested by the user
    (`/auth`, `/users`, `/vendors`, `/bookings`, `/payments`, `/execution`,
    `/notifications`, `/reviews`, `/search`, `/media`, `/chat`, `/ai`).

    Notes:
    - Most services return a common JSON envelope with `success`, `data`, and often `meta`.
    - Internal routes may accept either a bearer token with admin role or `x-internal-api-key`.
    - Chat real-time Socket.IO events are described in tag descriptions and README, while this spec focuses on HTTP endpoints.
servers:
  - url: http://localhost:8000
    description: Local API gateway
  - url: https://api-staging.weddingos.in
    description: Staging
  - url: https://api.weddingos.in
    description: Production
security:
  - bearerAuth: []
...
paths:
...
  /vendors/search:
    get:
      tags: [Vendors]
      summary: Search vendors from vendor-service index
      security: []
      parameters:
        - in: query
          name: q
          schema: { type: string }
        - in: query
          name: category
          schema: { type: string }
        - in: query
          name: city
          schema: { type: string }
        - in: query
          name: minRating
          schema: { type: number }
        - in: query
          name: minPricePaise
          schema: { type: integer }
        - in: query
          name: maxPricePaise
          schema: { type: integer }
        - in: query
          name: plusMembersOnly
          schema: { type: boolean }
        - in: query
          name: page
          schema: { type: integer, minimum: 1, default: 1 }
        - in: query
          name: limit
          schema: { type: integer, minimum: 1, maximum: 50, default: 20 }
        - in: query
          name: sortBy
          schema:
            type: string
            enum: [rating, price_asc, price_desc, bookings, relevance]
      responses:
        '200':
          description: Search result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VendorSearchEnvelope'
        '400':
          $ref: '#/components/responses/BadRequest'
  /vendors/suggest:
    get:
      tags: [Vendors]
      summary: Suggest vendors by prefix
      security: []
      parameters:
        - in: query
          name: q
          required: false
          schema: { type: string }
        - in: query
          name: city
          required: false
          schema: { type: string }
      responses:
        '200':
          description: Suggested vendor snippets
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      suggestions:
                        type: array
                        items:
                          $ref: '#/components/schemas/VendorSuggestion'
                  meta:
                    $ref: '#/components/schemas/ApiMeta'
  /vendors/{slug}:
    get:
      tags: [Vendors]
      summary: Get vendor by slug
      security: []
      parameters:
        - in: path
          name: slug
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Vendor detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VendorEnvelope'
        '404':
          $ref: '#/components/responses/NotFound'
  /vendors:
    post:
      tags: [Vendors]
      summary: Create vendor profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateVendorRequest'
      responses:
        '201':
          description: Created vendor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VendorEnvelope'
        '400':
          $ref: '#/components/responses/BadRequest'
  /vendors/me/profile:
    get:
      tags: [Vendors]
      summary: Get my vendor profile
      responses:
        '200':
          description: Vendor profile for authenticated vendor/admin
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VendorEnvelope'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      tags: [Vendors]
      summary: Update my vendor profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateVendorRequest'
      responses:
        '200':
          description: Updated vendor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VendorEnvelope'
        '404':
          $ref: '#/components/responses/NotFound'
  /vendors/me/packages:
    post:
      tags: [Vendors]
      summary: Create or update vendor package
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpsertPackageRequest'
      responses:
        '201':
          description: Upserted package
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VendorPackageEnvelope'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
  /vendors/me/packages/{packageId}:
    delete:
      tags: [Vendors]
      summary: Delete vendor package
      parameters:
        - in: path
          name: packageId
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: Package deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageEnvelope'
        '404':
          $ref: '#/components/responses/NotFound'
  /vendors/me/portfolio/presign:
    post:
      tags: [Vendors]
      summary: Get presigned portfolio upload URL
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                contentType:
                  type: string
                  example: image/jpeg
      responses:
        '200':
          description: Presigned upload data including vendorId
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    allOf:
                      - $ref: '#/components/schemas/PresignResult'
                      - type: object
                        properties:
                          vendorId:
                            type: string
                            format: uuid
                  meta:
                    $ref: '#/components/schemas/ApiMeta'
  /vendors/me/portfolio/confirm:
    post:
      tags: [Vendors]
      summary: Confirm uploaded portfolio item
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ConfirmPortfolioRequest'
      responses:
        '201':
          description: Portfolio item recorded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PortfolioItemEnvelope'
  /vendors/me/availability:
    post:
      tags: [Vendors]
      summary: Replace future availability blocks
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AvailabilityUpdateRequest'
      responses:
        '200':
          description: Availability updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageEnvelope'
...
  /vendors/admin/list:
    get:
      tags: [Vendors]
      summary: Admin list vendors
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Limit100'
        - in: query
          name: status
          schema: { type: string }
        - in: query
          name: category
          schema: { type: string }
        - in: query
          name: q
          schema: { type: string }
      responses:
        '200':
          description: Paginated vendor admin list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminVendorListEnvelope'
...
  /vendors/{id}/approve:
    patch:
      tags: [Vendors]
      summary: Approve vendor
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200':
          description: Approved vendor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VendorEnvelope'
  /vendors/{id}/reject:
    patch:
      tags: [Vendors]
      summary: Reject vendor
      parameters:
        - $ref: '#/components/parameters/IdPath'
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                reason:
                  type: string
      responses:
        '200':
          description: Rejected vendor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VendorEnvelope'
  /vendors/{id}/suspend:
    patch:
      tags: [Vendors]
      summary: Suspend vendor
      parameters:
        - $ref: '#/components/parameters/IdPath'
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                note:
                  type: string
      responses:
        '200':
          description: Suspended vendor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VendorEnvelope'
  /vendors/internal/{id}/bank-details:
    get:
      tags: [Vendors]
      summary: Get vendor bank details
      description: Internal/admin endpoint used by other services such as payment flows.
      security:
        - internalApiKey: []
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200':
          description: Vendor bank details
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean, example: true }
                  data:
                    type: object
                    properties:
                      vendor:
                        $ref: '#/components/schemas/VendorBankDetails'
                  meta:
                    $ref: '#/components/schemas/ApiMeta'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /reviews/vendor/{vendorId}:
    get:
      tags: [Reviews]
      summary: List published vendor reviews
      security: []
      parameters:
        - $ref: '#/components/parameters/VendorIdPath'
        - $ref: '#/components/parameters/Page'
        - in: query
          name: limit
          schema: { type: integer, minimum: 1, maximum: 50, default: 20 }
      responses:
        '200':
          description: Paginated vendor reviews and aggregate stats
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VendorReviewListEnvelope'
...
  /search/vendors:
    get:
      tags: [Search]
      summary: Search vendors from dedicated search-service
      security: []
      parameters:
        - in: query
          name: query
          schema: { type: string }
        - in: query
          name: category
          schema: { type: string }
        - in: query
          name: city
          schema: { type: string }
        - in: query
          name: minPrice
          schema: { type: number }
        - in: query
          name: maxPrice
          schema: { type: number }
        - in: query
          name: minRating
          schema: { type: number, minimum: 0, maximum: 5 }
        - in: query
          name: sortBy
          schema:
            type: string
            enum: [rating, price_asc, price_desc, reviews]
        - $ref: '#/components/parameters/Page'
        - in: query
          name: limit
          schema: { type: integer, minimum: 1, maximum: 50, default: 20 }
        - in: query
          name: featured
          schema: { type: boolean }
        - in: query
          name: lat
          schema: { type: number, minimum: -90, maximum: 90 }
        - in: query
          name: lng
          schema: { type: number, minimum: -180, maximum: 180 }
        - in: query
          name: radius
          schema: { type: number, exclusiveMinimum: 0 }
        - in: query
          name: pincode
          schema: { type: string }
      responses:
        '200':
          description: Search response with aggregations
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SearchVendorEnvelope'
        '400':
          $ref: '#/components/responses/BadRequest'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    internalApiKey:
      type: apiKey
      in: header
      name: x-internal-api-key
  parameters:
    Page:
      in: query
      name: page
      schema:
        type: integer
        minimum: 1
        default: 1
    Limit100:
      in: query
      name: limit
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    IdPath:
      in: path
      name: id
      required: true
      schema:
        type: string
        format: uuid
...
    VendorIdPath:
      in: path
      name: vendorId
      required: true
      schema:
        type: string
        format: uuid
...
  responses:
    BadRequest:
      description: Validation or malformed request error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
  schemas:
    ApiMeta:
      type: object
      properties:
        requestId:
          type: string
          nullable: true
        timestamp:
          type: string
          format: date-time
...
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          $ref: '#/components/schemas/ErrorObject'
        meta:
          $ref: '#/components/schemas/ApiMeta'
    MessageEnvelope:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
          properties:
            message:
              type: string
        meta:
          $ref: '#/components/schemas/ApiMeta'
...
    PackageType:
      type: string
      enum: [BASIC, STANDARD, PREMIUM, CUSTOM]
...
    PresignResult:
      type: object
      properties:
        uploadUrl: { type: string }
        key: { type: string }
        s3Key: { type: string, nullable: true }
        publicUrl: { type: string }
        expiresIn: { type: integer, example: 300 }
        mock: { type: boolean, nullable: true }
...
    CreateVendorRequest:
      type: object
      required: [businessName, category, city, state, pincode]
      properties:
        businessName: { type: string, minLength: 2, maxLength: 200 }
        category: { type: string }
        city: { type: string, maxLength: 100 }
        state: { type: string, maxLength: 100 }
        pincode:
          type: string
          pattern: '^\d{6}$'
    UpdateVendorRequest:
      type: object
      properties:
        businessName: { type: string, maxLength: 200 }
        tagline: { type: string, maxLength: 300 }
        description: { type: string, maxLength: 5000 }
        city: { type: string, maxLength: 100 }
        state: { type: string, maxLength: 100 }
        pincode:
          type: string
          pattern: '^\d{6}$'
        serviceCities:
          type: array
          items: { type: string }
        yearsExperience: { type: integer, minimum: 0, maximum: 100 }
        teamSize: { type: integer, minimum: 1, maximum: 10000 }
        whatsappNumber: { type: string }
        websiteUrl: { type: string, format: uri }
        instagramUrl: { type: string, format: uri }
        gstNumber: { type: string }
        panNumber: { type: string }
        bankAccountNo: { type: string }
        bankIfsc: { type: string }
        bankAccountName: { type: string }
    VendorPackage:
      type: object
      properties:
        id: { type: string, format: uuid }
        vendorId: { type: string, format: uuid }
        packageType:
          $ref: '#/components/schemas/PackageType'
        name: { type: string }
        description: { type: string, nullable: true }
        priceFromPaise: { type: integer }
        priceUpToPaise: { type: integer, nullable: true }
        isCustomQuote: { type: boolean }
        inclusions:
          type: array
          items: { type: string }
        exclusions:
          type: array
          items: { type: string }
        deliverables:
          type: array
          items: { type: string }
        isActive: { type: boolean }
        sortOrder: { type: integer }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    UpsertPackageRequest:
      type: object
      required: [name, packageType, priceFromPaise]
      properties:
        id: { type: string, format: uuid }
        name: { type: string, minLength: 2, maxLength: 200 }
        packageType:
          $ref: '#/components/schemas/PackageType'
        priceFromPaise: { type: integer, minimum: 0 }
        priceUpToPaise: { type: integer }
        isCustomQuote: { type: boolean }
        description: { type: string, maxLength: 2000 }
        inclusions:
          type: array
          items: { type: string }
        exclusions:
          type: array
          items: { type: string }
        deliverables:
          type: array
          items: { type: string }
    PortfolioItem:
      type: object
      properties:
        id: { type: string, format: uuid }
        vendorId: { type: string, format: uuid }
        s3Key: { type: string }
        publicUrl: { type: string }
        thumbUrl: { type: string, nullable: true }
        caption: { type: string, nullable: true }
        mediaType: { type: string }
        sortOrder: { type: integer }
        createdAt: { type: string, format: date-time }
...
    VendorEnvelope:
      type: object
      properties:
        success: { type: boolean, example: true }
        data:
          type: object
          properties:
            vendor:
              $ref: '#/components/schemas/Vendor'
        meta:
          $ref: '#/components/schemas/ApiMeta'
    VendorPackageEnvelope:
      type: object
      properties:
        success: { type: boolean, example: true }
        data:
          type: object
          properties:
            package:
              $ref: '#/components/schemas/VendorPackage'
        meta:
          $ref: '#/components/schemas/ApiMeta'
    PortfolioItemEnvelope:
      type: object
      properties:
        success: { type: boolean, example: true }
        data:
          type: object
          properties:
            item:
              $ref: '#/components/schemas/PortfolioItem'
        meta:
          $ref: '#/components/schemas/ApiMeta'
    VendorSuggestion:
      type: object
      properties:
        id: { type: string }
        businessName: { type: string }
        slug: { type: string }
        category: { type: string }
        city: { type: string }
        avgRating: { type: number, nullable: true }
        coverPhoto: { type: string, nullable: true }
    VendorSearchResult:
      type: object
      properties:
        vendors:
          type: array
          items:
            $ref: '#/components/schemas/Vendor'
        total: { type: integer }
        page: { type: integer }
        limit: { type: integer }
        totalPages: { type: integer }
    VendorSearchEnvelope:
      type: object
      properties:
        success: { type: boolean, example: true }
        data:
          $ref: '#/components/schemas/VendorSearchResult'
        meta:
          $ref: '#/components/schemas/ApiMeta'
    ConfirmPortfolioRequest:
      type: object
      required: [s3Key, publicUrl]
      properties:
        s3Key: { type: string }
        publicUrl: { type: string }
        caption: { type: string }
        mediaType: { type: string }
    AvailabilityUpdateRequest:
      type: object
      properties:
        blockedDates:
          type: array
          items:
            type: object
            required: [date]
            properties:
              date:
                type: string
                pattern: '^\d{4}-\d{2}-\d{2}$'
              reason:
                type: string
    VendorBankDetails:
      type: object
      properties:
        id: { type: string, format: uuid }
        bankAccountNo: { type: string, nullable: true }
        bankIfsc: { type: string, nullable: true }
        bankAccountName: { type: string, nullable: true }
    AdminVendorListEnvelope:
      type: object
      properties:
        success: { type: boolean, example: true }
        data:
          type: object
          properties:
            vendors:
              type: array
              items:
                $ref: '#/components/schemas/Vendor'
        meta:
          allOf:
            - $ref: '#/components/schemas/ApiMeta'
            - type: object
              properties:
                total: { type: integer }
                page: { type: integer }
                limit: { type: integer }
                pages: { type: integer }
...
    VendorReviewStats:
      type: object
      properties:
        _avg:
          type: object
          properties:
            rating: { type: number, nullable: true }
            qualityRating: { type: number, nullable: true }
            valueRating: { type: number, nullable: true }
            professionalismRating: { type: number, nullable: true }
            punctualityRating: { type: number, nullable: true }
        _count:
          type: object
          properties:
            id: { type: integer }
    VendorReviewListEnvelope:
      type: object
      properties:
        success: { type: boolean, example: true }
        data:
          type: object
          properties:
            reviews:
              type: array
              items:
                $ref: '#/components/schemas/Review'
            total: { type: integer }
            page: { type: integer }
            limit: { type: integer }
            totalPages: { type: integer }
            stats:
              $ref: '#/components/schemas/VendorReviewStats'
...
    SearchVendorEnvelope:
      type: object
      properties:
        success: { type: boolean, example: true }
        data:
          type: object
          properties:
            vendors:
              type: array
              items:
                $ref: '#/components/schemas/SearchVendorHit'
            total: { type: integer }
            page: { type: integer }
            limit: { type: integer }
            totalPages: { type: integer }
            aggregations:
              $ref: '#/components/schemas/SearchAggregations'
...